+++ .build.ps1
@@ -159,10 +159,10 @@ task Test {
 	$output = Invoke-Build . Demo\.build.ps1 -Result result | Out-String -Width:9999
 	if ($SkipTestDiff) { return }
 
-	assert (167 -eq $result.AllTasks.Count) $result.AllTasks.Count
+	assert (174 -eq $result.AllTasks.Count) $result.AllTasks.Count
 	assert (32 -eq $result.Tasks.Count) $result.Tasks.Count
 
-	assert (30 -eq $result.AllErrorCount) $result.AllErrorCount
+	assert (34 -eq $result.AllErrorCount) $result.AllErrorCount
 	assert (0 -eq $result.ErrorCount) $result.AllErrorCount
 
 	assert ($result.AllWarningCount -ge 1)