+++ Demo/Parallel.build.ps1
@@ -7,6 +7,8 @@
 	Invoke-Build * Parallel.build.ps1
 #>
 
+. .\SharedScript.ps1
+
 # The build engine defines aliases Invoke-Build and Invoke-Builds
 task Alias {
 	#! also covers 1.2.0, see notes
@@ -53,7 +55,7 @@ techniques useful for advanced build result analysis.
 task Many {
 	# These build parameter hashes are not changed (to be tested).
 	$build0 = @{File='Dynamic.build.ps1'}
-	$build1 = @{File='ErrorCases.build.ps1'}
+	$build1 = @{File='ErrorCases.build.ps1'; Task='Error1'}
 	$build2 = @{File='Dynamic.build.ps1'; Task='Task0'}
 	$build3 = @{File='Conditional.build.ps1'; Parameters=@{Configuration='Debug'}}
 
@@ -67,7 +69,7 @@ task Many {
 	# Get the results using the variable name.
 	$message = ''
 	try {
-		Invoke-Builds -Build $build -Result Result #@{Configuration='Release'}
+		Invoke-Builds -Build $build -Result Result
 	}
 	catch {
 		$message = "$_"
@@ -87,7 +89,7 @@ task Many {
 	assert ($build[0].Result.Value)
 
 	# The call itself failed.
-	assert ($message -like "Parallel build failures:*Build: *\ErrorCases.build.ps1*ERROR: Error1*At *\SharedTasksData.tasks.ps1:*")
+	assert ($message -like "Failed builds:*Build: *\ErrorCases.build.ps1*ERROR: Error1*At *\SharedTasksData.tasks.ps1:*")
 
 	# Check each build error; note that three builds succeeded because the
 	# parallel build engine lets all builds to work, even if some fail.
@@ -103,9 +105,14 @@ task Many {
 	assert (1 -eq $build[3].Result.Value.Tasks.Count)
 }
 
-# Invoke three builds with specified timeout. One build should complete, the
-# other two should be stopped. NB: Skip this task in problematic locations.
-task Timeout -If (!$BuildRoot.Contains('[')) {
+<#
+Invoke three builds with the specified timeout. The first build should
+complete, the other two should be stopped by the engine due to timeout.
+
+Skip this task in problem paths:
+https://connect.microsoft.com/PowerShell/feedback/details/708432
+#>
+task Timeout -If ($BuildRoot -notmatch '[\[\]]') {
 	# Invoke using try..catch because it fails and use log files for outputs.
 	$message = ''
 	try {
@@ -121,7 +128,7 @@ task Timeout -If (!$BuildRoot.Contains('[')) {
 
 	# Check the error message.
 	assert ($message -like @'
-Parallel build failures:
+Failed builds:
 Build: *\Sleep.build.ps1
 ERROR: Build timed out.
 Build: *\Sleep.build.ps1
@@ -133,4 +140,50 @@ ERROR: Build timed out.*
 	assert (!(Test-Path z.2) -or (Get-Content z.2)[-1] -eq 'begin - начало')
 	assert (!(Test-Path z.2) -or (Get-Content z.3)[-1] -eq 'begin - начало')
 	Remove-Item z.?
+}
+
+# Error: invalid MaximumBuilds
+task ParallelBadMaximumBuilds {
+	Invoke-Builds -MaximumBuilds 0 @{File='Dynamic.build.ps1'}
+}
+
+# Error: Invoke-Build.ps1 is not in the same directory.
+task ParallelMissingEngine {
+	try {
+		Copy-Item -LiteralPath (Get-Alias Invoke-Builds).Definition $env:TEMP
+		& "$env:TEMP\Invoke-Builds.ps1" @{File='Sleep.build.ps1'}
+	}
+	finally {
+		[System.IO.File]::Delete("$env:TEMP\Invoke-Builds.ps1")
+	}
+}
+
+# Error: missing mandatory File
+task ParallelNoFileParameter {
+	Invoke-Builds @{Task='.'}
+}
+
+# Error: missing build File
+task ParallelMissingFile {
+	Invoke-Builds @{File='MissingFile'}
+}
+
+# Error: invalid Parameters type on calling Invoke-Builds
+task ParallelBadParameters {
+	Invoke-Builds @{File='Dynamic.build.ps1'; Parameters='BadParameters'}
+}
+
+# Test error cases.
+task ParallelErrorCases `
+@{ParallelMissingEngine=1},
+@{ParallelNoFileParameter=1},
+@{ParallelMissingFile=1},
+@{ParallelBadMaximumBuilds=1},
+@{ParallelBadParameters=1},
+{
+	Test-Error ParallelMissingEngine "Required script '*\Invoke-Build.ps1' does not exist.*At *\Parallel.build.ps1:*ObjectNotFound*"
+	Test-Error ParallelNoFileParameter "Build parameter File is missing or empty.*Invoke-Builds <<<<*InvalidArgument*"
+	Test-Error ParallelMissingFile "Build file '*\MissingFile' does not exist.*Invoke-Builds <<<<*ObjectNotFound*"
+	Test-Error ParallelBadMaximumBuilds "MaximumBuilds should be a positive number.*Invoke-Builds <<<<*InvalidArgument*"
+	Test-Error ParallelBadParameters "Failed builds:*Build: *\Dynamic.build.ps1*ERROR: '*\Dynamic.build.ps1' invocation failed:*"
 }